move helper.py from as7726-32X to common
